555定时器-PWM Input on CH1模式测频率和占空比

+++ Core/Inc/app.h
@@ -23,7 +23,7 @@ typedef struct {
 typedef struct {
     uint32_t pa6_freq;
     float pa15_freq;
-    uint32_t pa15_duty;
+    float pa15_duty;
 } Measure_t;
 
 void App_Init(void);
